better?

+++ EmbeddedTesting/motor_test.py
@@ -42,7 +42,7 @@ while True:
         mymotortest.motor_go(False, # True=Clockwise, False=Counter-Clockwise
                             "Full" , # Step type (Full,Half,1/4,1/8,1/16,1/32)
                             200, # number of steps
-                            int(command), # step delay [sec]
+                            float(command), # step delay [sec]
                             False, # True = print verbose output 
                             .05) # initial delay [sec]
